pyproject.toml
[project]
name = "iconforge"
version = "0.1.0"
description = "A tool for icon processing and manipulation"
requires-python = ">=3.8"
dependencies = [
    "numpy==2.2.6",
    "onnxruntime-gpu==1.22.0",
    "opencv-python==4.11.0.86",
    "pillow==11.2.1",
    "protobuf==6.31.1",
    "python-dotenv==1.0.1",
    "rembg==2.0.66",
    "requests==2.32.4",
    "safetensors==0.5.3",
    "scikit-image==0.25.2",
    "tqdm==4.67.1",
    "transparent-background==1.3.4",
    "typing_extensions==4.12.2",
    "urllib3==2.4.0",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]
